fix: move ArtworkThumbnail out of InfiniteDiscovery into shared Components
BottomTabs was reaching into InfiniteDiscovery for this component, which
AGENTS.md forbids importing across Scenes. It has no InfiniteDiscovery-
specific dependencies, so it moves to app/Components without changes.
