+++ src/app/Components/ArtworkThumbnail.tsx
